Skip update content if CONTENT_DIR is "content".

+++ bin/build.sh
@@ -2,7 +2,10 @@
 
 set -eu
 
-rm -rf content
-mkdir content
-cp -r ${CONTENT_DIR}/* content
+# Skip build if CONTENT_DIR is not content
+if [ "${CONTENT_DIR}" != "content" ]; then
+  rm -rf content
+  mkdir content
+  cp -r ${CONTENT_DIR}/* content
+fi
 npx quartz build $@
